+++ internal/cli/options.go
@@ -1,50 +1,65 @@
 package cli
 
-const (
+// Option is a type representing a CLI argument option or flag.
+type Option struct {
+	// The full version of the Option.
+	name string
+
+	// The alias for the Option.
+	alias string
+}
+
+var (
 	// The flag to output the usage of the program.
-	helpFlag = "--help"
+	helpFlag = Option{
+		name: "--help",
+	}
 
 	// The flag to output the program version.
-	versionFlag = "--version"
+	versionFlag = Option{
+		name: "--version",
+	}
 
-	// The flag to enable dry run. If enabled the program won't make any changes to
-	// the input files.
-	dryRunFlag = "--dry-run"
+	// The flag to enable dry run. If enabled the program won't make any changes
+	// to the input files.
+	dryRunFlag = Option{
+		name: "--dry-run",
+	}
 
 	// The flag to invert the mapping. If enabled the mapping will be used right-
 	// to-left instead of left-to-right.
-	invertFlag = "--invert"
-
-	// The alias for the --invert flag.
-	invertFlagAlias = "-i"
+	invertFlag = Option{
+		name:  "--invert",
+		alias: "-i",
+	}
 
 	// The flag to make the program silent.
-	silentFlag = "--silent"
-
-	// The alias for the --silent flag.
-	silentFlagAlias = "-s"
+	silentFlag = Option{
+		name:  "--silent",
+		alias: "-s",
+	}
 
 	// The flag to make the program verbose.
-	verboseFlag = "--verbose"
-
-	// The alias for the --verbose flag.
-	verboseFlagAlias = "-v"
+	verboseFlag = Option{
+		name:  "--verbose",
+		alias: "-v",
+	}
 
 	// The option to specify the configuration file to use.
-	configOption = "--config"
-
-	// The alias for the --config option.
-	configOptionAlias = "-c"
+	configOption = Option{
+		name:  "--config",
+		alias: "-c",
+	}
 
 	// The option to specify a mapping file.
-	mapfileOption = "--map-file"
-
-	// The alias for the --map-file option.
-	mapfileOptionAlias = "-M"
+	mapfileOption = Option{
+		name:  "--map-file",
+		alias: "-M",
+	}
 
 	// The option to specify a single mapping from the CLI.
-	mappingOption = "--map"
-
-	// The alias for the --map option.
-	mappingOptionAlias = "-m"
+	mappingOption = Option{
+		name:  "--map",
+		alias: "-m",
+	}
 )