+++ internal/wordmaps/markdown_fuzz.go
@@ -1,0 +1,8 @@
+package wordmaps
+
+// FuzzMarkDown will fuzz the MarkDown parser.
+func FuzzMarkDown(data []byte) int {
+	s := string(data)
+	parseMarkDownFile(&s)
+	return 0
+}